Dzialajace dodawanie punktow

+++ src/main/java/com/example/Main/Model/Scores.java
@@ -15,7 +15,7 @@ public class Scores
     private String username;
     @Basic
     @Column(name = "score")
-    private Integer score;
+    private Double score;
 
     public int getUid()
     {
@@ -37,12 +37,12 @@ public class Scores
         this.username = username;
     }
 
-    public Integer getScore()
+    public Double getScore()
     {
         return score;
     }
 
-    public void setScore(Integer score)
+    public void setScore(Double score)
     {
         this.score = score;
     }
